Split browser.py
